i40pcf: Fix errors with not creating proper assets with double entries
Catalog broken in cases a contract definition is created with a null
asset. Therefore, make list of endpoints unique entries and
also check in case asset could not be created, do NOT continue
with policy and contract definition.

Also proper output to show what has been created and what not.

Signed-off-by: Matthias Binzer <matthias.binzer@de.bosch.com>