auth: First draft with an Auth Server via EDC
Instead of a data source behind the Dataplane,
the call goes to an Authorization Server that
creates an access_token that can be used in the
next call to access the actual data.

Background: The goal is to use more than just the
EDR token in the token exchange (/token),
e.g. some additional VCs

Signed-off-by: Matthias Binzer <matthias.binzer@de.bosch.com>